+++ Source/BugsnagUser.h
@@ -15,8 +15,8 @@
 
 - (NSDictionary *)toJson;
 
-@property NSString *userId;
-@property NSString *name;
-@property NSString *emailAddress;
+@property(readonly) NSString *userId;
+@property(readonly) NSString *name;
+@property(readonly) NSString *emailAddress;
 
 @end