+++ CHANGELOG.md
@@ -11,6 +11,12 @@ Bugsnag Notifiers on other platforms.
 * Make user editable on `BugsnagEvent`
   [#557](https://github.com/bugsnag/bugsnag-cocoa/pull/557)
 
+* Add getter for `Bugsnag.context`
+  [#554](https://github.com/bugsnag/bugsnag-cocoa/pull/554)
+
+* Make `BugsnagUser` properties readonly
+  [#556](https://github.com/bugsnag/bugsnag-cocoa/pull/556)
+
 * Add `sendThreads` property to `BugsnagConfiguration`
   [#549](https://github.com/bugsnag/bugsnag-cocoa/pull/549)